- Implemented client side routing using react-router v4 library. - Added skeleton routes and components for the following routes: /: homepage /builder: analysis builder interface /browse: public analysis browser - Added first draft of frontend dev notes (DEVNOTES.md).
