+++ web/frontend/src/App.css
@@ -24,3 +24,7 @@
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
+
+body {
+    background-color: white;
+}
